% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{capture_error}
\alias{capture_error}
\title{Capture errors in code execution}
\usage{
capture_error(code, otherwise = NULL, quiet = TRUE)
}
\arguments{
\item{code}{The code to be executed.}

\item{otherwise}{A default value to return if an error occurs. Default is \code{NULL}.}

\item{quiet}{A logical value indicating whether to suppress error messages.
Default is \code{TRUE}.}
}
\value{
A list with two elements: \code{result} (the result of the code execution)
and \code{error} (the error that occurred, if any).
}
\description{
This function executes code and captures any errors that occur, returning a list with the
result and the error.
}
