% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_filter_texts.R
\name{filter_for_keywords}
\alias{filter_for_keywords}
\title{Filter a tibble for keywords based on the YAML config}
\usage{
filter_for_keywords(df, config)
}
\arguments{
\item{df}{A tibble containing rss data}

\item{config}{YAML config
\itemize{
  \item \code{rss_filter$keywords}: Character vector of keywords to search for
  \item \code{rss_filter$search_in}: Character vector of column names to search within ('item_text' and/or 'item_title')
}}
}
\value{
A filtered tibble where the specified columns contain one of the keywords (case-insensitive)
}
\description{
Filter a tibble for keywords based on the YAML config
}
