% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_process_yaml_config.R
\name{load_yaml_config}
\alias{load_yaml_config}
\title{Load YAML Configuration}
\usage{
load_yaml_config(config_file = "config.yaml")
}
\arguments{
\item{config_file}{(str) name of the config file}
}
\value{
(list) The parsed YAML content
}
\description{
Load YAML Configuration
}
