% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_logging.R
\name{logging}
\alias{logging}
\title{Print logging messages}
\usage{
logging(msg, ...)
}
\arguments{
\item{msg}{message to log}

\item{...}{further params passed to flog.info}
}
\description{
Print logging messages
}
