% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_process_yaml_config.R
\name{read_field_mapping}
\alias{read_field_mapping}
\title{Read Mapping Table for DB and RSS from YAML Config}
\usage{
read_field_mapping(config_file = "config.yaml")
}
\arguments{
\item{config_file}{(str) name of config file}
}
\description{
Read Mapping Table for DB and RSS from YAML Config
}
