% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_process_yaml_config.R
\name{read_rss_fields}
\alias{read_rss_fields}
\title{Read RSS field names for text and url to text from Configuration
Needed in case of changes to the mapping table (e.g. renaming of database columns)}
\usage{
read_rss_fields(config)
}
\arguments{
\item{config}{(list) the configuration object}
}
\description{
Read RSS field names for text and url to text from Configuration
Needed in case of changes to the mapping table (e.g. renaming of database columns)
}
