% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_process_yaml_config.R
\name{read_rss_sources}
\alias{read_rss_sources}
\title{Read RSS Sources from YAML Config}
\usage{
read_rss_sources(config_file = "config.yaml")
}
\arguments{
\item{config_file}{(str) name of config file}
}
\description{
Read RSS Sources from YAML Config
}
