% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{read_text_via_url}
\alias{read_text_via_url}
\title{Read item_text via Provided Link in RSS Feed}
\usage{
read_text_via_url(url)
}
\arguments{
\item{url}{(str) item_link from the rss feed}
}
\description{
Read item_text via Provided Link in RSS Feed
}
