% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{safe_GET}
\alias{safe_GET}
\title{Safely Wrapped GET Function from httr}
\usage{
safe_GET(...)
}
\arguments{
\item{...}{Additional arguments passed to the \code{httr::HEAD} function.}
}
\value{
A function that performs a GET request and returns a list with elements
\code{result} and \code{error}.
}
\description{
This function is a safely wrapped version of \code{httr::GET} using the \code{safely} function.
It captures errors and returns a list with the result and the error.
}
