% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{safe_HEAD}
\alias{safe_HEAD}
\title{Safely Wrapped HEAD Function from httr}
\usage{
safe_HEAD(...)
}
\arguments{
\item{...}{Additional arguments passed to the \code{httr::HEAD} function.}
}
\value{
A function that performs a HEAD request and returns a list with elements
\code{result} and \code{error}.
}
\description{
This function is a safely wrapped version of \code{httr::HEAD} using the \code{safely} function.
It captures errors and returns a list with the result and the error.
}
