% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_sync_sources.R
\name{sync_sources_with_mongo}
\alias{sync_sources_with_mongo}
\title{Sync config sources with MongoDB}
\usage{
sync_sources_with_mongo(config)
}
\arguments{
\item{config}{The loaded configuration list (as from config.yaml)}
}
\description{
Compares the sources listed in the config object
with the MongoDB "sources" collection.
Adds new sources, marks removed ones, and if a source with an existing
ID has changed (name or URL), it raises a warning, marks the old entry
as removed, and inserts the updated version.
}
