% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{url_exists}
\alias{url_exists}
\title{Check if a URL exists and is reachable}
\usage{
url_exists(url, quiet = FALSE, ...)
}
\arguments{
\item{url}{A string representing the URL to check.}

\item{quiet}{A logical value indicating whether to suppress error messages.
Default is \code{FALSE}.}

\item{...}{Additional arguments passed to the \code{httr::HEAD} function.}
}
\value{
\code{TRUE} if the URL exists and is reachable, \code{FALSE} if the URL is not
reachable, and logs a warning message.
}
\description{
This function performs a HEAD request to check if a URL exists and is reachable.
}
