% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{read_rss_feeds}
\alias{read_rss_feeds}
\title{Read RSS Feeds}
\usage{
read_rss_feeds(feed, fields = read_field_mapping()$rss_names)
}
\arguments{
\item{fields}{(str) fields to read from the rss feed}

\item{feed_url}{(str) url of the rss feed}
}
\description{
Read RSS Feeds
}
