% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_read_rss.R
\name{read_rss_feeds}
\alias{read_rss_feeds}
\title{Read RSS Feeds}
\usage{
read_rss_feeds(source, config)
}
\arguments{
\item{source}{(list) source feed entry from the configuration file.}

\item{config}{(list) configuration file with column mapping and rss field names.}
}
\value{
A data frame containing the requested fields from the RSS feed.
}
\description{
This function reads and processes RSS feeds, extracting specified fields.
}
