% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_process_yaml_config.R
\name{read_rss_sources}
\alias{read_rss_sources}
\title{Read RSS Sources from Configuration}
\usage{
read_rss_sources(config)
}
\arguments{
\item{config}{(list) the configuration object}
}
\description{
Read RSS Sources from Configuration
}
