% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_db_creds.R
\name{update_database}
\alias{update_database}
\title{Update Database with New RSS Feed Data}
\usage{
update_database(tibble, config)
}
\arguments{
\item{tibble}{A tibble containing RSS feed data with unique `source_id`
and `feed_last_build_date`.}

\item{config}{(list) configuration file with column mapping and rss field names.}
}
\description{
This function updates the database with new RSS feed data from a list of tibbles.
For each tibble, it compares the `feed_last_build_date` with the corresponding
date in the database and updates the database if the tibble contains newer data.
}
